feat(css-manager): css manager and refactor css variables
