+++ sdks/web-sdk/src/lib/atoms/Input/Input.svelte
@@ -5,7 +5,7 @@
     ICSSProperties,
     IElementProps,
   } from '../../contexts/configuration';
-  import { makeStylesFromConfiguration } from '../../utils/css-utils';
+  import { makeStylesFromConfiguration } from '../../services/css-manager';
   import { IInputAttributes } from './types';
   import { t } from '../../contexts/translation/hooks';
   import merge from 'deepmerge';
