+++ sdks/web-sdk/src/lib/atoms/ListItem/ListItem.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { ICSSProperties, IElementProps } from '../../contexts/configuration';
-  import { makesLocalStyles } from '../../utils/css-utils';
+  import { makesLocalStyles } from '../../services/css-manager';
 
   export let configuration: IElementProps;
   const styleProps = configuration.style as ICSSProperties;
