+++ sdks/web-sdk/src/lib/contexts/configuration/types.ts
@@ -5,6 +5,7 @@ import type { CaptureConfigOption } from 'jslib-html5-camera-photo';
 import { IFlow } from '../flows';
 import { EndUserInfo, FlowsBackendConfig } from '../../../types/BallerineSDK';
 import { DocumentType } from '../app-state';
+import { ICSSProperties } from '../../services/css-manager';
 
 export enum Steps {
   Welcome = 'welcome',
@@ -46,48 +47,6 @@ export enum Elements {
   ErrorText = 'error-text',
 }
 
-export interface IOpacityColor {
-  color: string;
-  opacity: number;
-}
-
-interface ICSSPropertiesAll {
-  margin: string;
-  padding: string;
-  'font-family': string;
-  'font-size': string;
-  'font-weight': number;
-  'text-align': string;
-  'line-height': string;
-  'vertical-align': string;
-  'box-shadow': string;
-  '-webkit-box-shadow': string;
-  width: string;
-  height: string;
-  background?: string | IOpacityColor;
-  color?: string;
-  'border-radius'?: string;
-  border: string;
-  display: string;
-  cursor: string;
-  'align-items': string;
-  'justify-content': string;
-  'flex-direction': string;
-  position: string;
-  top: string;
-  bottom: string;
-  left: string;
-  right: string;
-  'align-self': string;
-  hover: ICSSProperties;
-  active: ICSSProperties;
-  fill: string;
-  'flex-grow': number;
-  'background-position-y': string;
-  outline: string;
-  'z-index': string;
-}
-
 export type Icons = 'Card' | 'Passport' | 'License' | 'PassportTwo' | 'Voter';
 
 interface IAttributesAll {
@@ -104,8 +63,6 @@ interface IAttributesAll {
   validate: () => boolean;
   defaultValue: string;
 }
-
-export type ICSSProperties = Partial<ICSSPropertiesAll>;
 export type IAttributes = Partial<IAttributesAll>;
 
 export interface IFormProps {
