+++ sdks/web-sdk/src/lib/services/css-manager/index.ts
@@ -1,0 +1,2 @@
+export type { ICSSProperties } from "./types";
+export { makeStylesFromConfiguration, makesLocalStyles } from "./utils";
